Use base_opt abstraction in base_opt and refactor existing minimize functions
