+++ skopt/optimizer/__init__.py
@@ -1,0 +1,4 @@
+from . dummy_opt import dummy_minimize
+from . gbrt_opt import gbrt_minimize
+from . forest_opt import forest_minimize
+from . gp_opt import gp_minimize
